adding delete route & updating form.js

+++ server/models/Item.js
@@ -2,7 +2,7 @@ let { db, DataTypes } = require('./connection.js');
 
 let Item = db.define('Item', {
     title: DataTypes.STRING,
-    price: DataTypes.INTEGER,
+    price: DataTypes.FLOAT, // Use FLOAT to support decimal values
     description: DataTypes.STRING,
     category: DataTypes.STRING,
     image: DataTypes.STRING,
